+++ frontend/src/components/index.ts
@@ -1,4 +1,4 @@
-export { default as Header } from './Header';
+export { default as LayoutNavigation } from './LayoutNavigation';
 export { default as Input } from './Input';
 export { default as MainCardList } from './MainCardList';
 export { default as SaleSchedule } from './SaleSchedule';
