+++ frontend/src/pages/Home/index.tsx
@@ -1,4 +1,4 @@
-import { Input, MainCardList, Header, CardSlider } from '../../components';
+import { Input, MainCardList, LayoutNavigation, CardSlider } from '../../components';
 import styled from 'styled-components';
 import PALETTE from '../../constants/palette';
 import { SubscriptionUsedMainPage } from '../../types';
@@ -39,7 +39,7 @@ const Home = () => {
   console.log(likeList);
 
   return (
-    <Header>
+    <LayoutNavigation>
       <Input placeholder="검색어를 입력하세요 (ex.지역)" />
       <StyledMainWrapper>
         <div>오늘의 청약</div>
@@ -49,7 +49,7 @@ const Home = () => {
       </StyledMainWrapper>
       <CardSlider todaySubscriptions={todaySubscriptions} />
       <MainCardList popularityList={popularityList} likeList={likeList} />
-    </Header>
+    </LayoutNavigation>
   );
 };
